+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -321,7 +321,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSSSKVK94VWN12YJ25W4FW68"
+        "01KSSXNJ2M9ZFMGABHVKTJGBQ1"
     )
     r"""The unique identifier of the data source"""