+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSSSKVK782F2AMNTNX916S5T"
+        "01KSSXNJ2J7XS66G8PTYD17Y09"
     )
     r"""The unique identifier of the data source"""