+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSSSKVK8CECCPZKS0G2Y0S2M"
+        "01KSSXNJ2MZ7E1V0Y2J8VJ4N8K"
     )
     r"""The unique identifier of the data source"""