+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KSSSKVKB2JNTQ5FC89J9RFKC"
+        "01KSSXNJ2P3W4MEJ0J68ZA82DB"
     )
     r"""The unique identifier of the data source"""