+++ packages/orq-rc/RELEASES.md
@@ -8299,3 +8299,13 @@ Based on:
 - [python v4.10.0-rc.56] packages/orq-rc
 ### Releases
 - [PyPI v4.10.0-rc.56] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.56 - packages/orq-rc
+
+## 2026-05-29 13:12:33
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.768.2 (2.889.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.57] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.57] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.57 - packages/orq-rc